third_party/blink/web_tests/virtual/print_browser/print_testharness/test_enable_print_browser.html

<!DOCTYPE html>
<html>

<style style="text/css">
  @media screen
  {
    p#bodyText {color: red; font-size: 20px;}
  }
  @media print
  {
    p#bodyText {color: green; font-size: 40px;}
  }
</style>

<body>
  <p id="bodyText">Test message please ignore.</p>
</body>

<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script>
var t = async_test();
// The <style> is guaranteed to be set by the time this script runs, but print
// mode is set via an IPC which races with the loading task which runs this
// script. We get @media screen until that IPC is processed.
window.onload = function() {
  t.step(function () {
    assert_equals(getComputedStyle(bodyText).fontSize, "40px")
  }, "Checks if we are in print mode");
  t.done();
};
</script>
</html>
